2009-11-13  Chris Sutcliffe  <ir0nh34d@users.sourceforge.net>
        * include/io.h (_open_osfhandle): Correct definition.

        Thanks to Alexander Shaduri for the information.

+++ winsup/mingw/include/io.h
@@ -238,7 +238,7 @@ _CRTIMP long __cdecl __MINGW_NOTHROW _lseek (int, long, int);
 /* Optional third argument is unsigned unPermissions. */
 _CRTIMP int __cdecl __MINGW_NOTHROW _open (const char*, int, ...);
 
-_CRTIMP int __cdecl __MINGW_NOTHROW _open_osfhandle (long, int);
+_CRTIMP int __cdecl __MINGW_NOTHROW _open_osfhandle (intptr_t, int);
 _CRTIMP int __cdecl __MINGW_NOTHROW _pipe (int *, unsigned int, int);
 _CRTIMP int __cdecl __MINGW_NOTHROW _read (int, void*, unsigned int);
 _CRTIMP int __cdecl __MINGW_NOTHROW _setmode (int, int);
